remaining crl hours as float

+++ encryptme_stats/metrics.py
@@ -403,7 +403,7 @@ def openssl():
             'openssl':  {
                 'crl_last_update': last_update.isoformat(),
                 'crl_next_update': next_update.isoformat(),
-                'crl_remaining_hours': int((next_update - now).total_seconds() / 3600.0),
+                'crl_remaining_hours': float((next_update - now).total_seconds() / 3600.0),
                 'certificate_start_date': start_date.isoformat(),
                 'certificate_end_date': end_date.isoformat(),
                 'certificate_remaining_days': (end_date - now).days
